Move dnsxd app out of dnsxd subfolder
R15B01's reltool has an app-level lib_dir setting which makes it no longer necessary for the dnsxd
app to reside in a subfolder. Taking advantage of this makes it much easier to reuse dnsxd as a
dependancy of another app although it does break release generation for earlier releases.